+++ ethergo/submitter/submitter.go
@@ -4,16 +4,15 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"github.com/ethereum/go-ethereum/params"
-	"github.com/puzpuzpuz/xsync/v2"
-	"github.com/synapsecns/sanguine/ethergo/chain/gas/backend"
-	"github.com/synapsecns/sanguine/ethergo/chain/gas/london"
 	"math"
 	"math/big"
 	"reflect"
 	"runtime"
 	"sync"
 	"time"
+
+	"github.com/google/uuid"
+	"github.com/puzpuzpuz/xsync/v2"
 
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
@@ -214,15 +213,16 @@ func (t *txSubmitterImpl) getNonce(parentCtx context.Context, chainID *big.Int, 
 	return dbNonce, nil
 }
 
-func (t *txSubmitterImpl) storeTX(ctx context.Context, tx *types.Transaction, status db.Status) (err error) {
+func (t *txSubmitterImpl) storeTX(ctx context.Context, tx *types.Transaction, status db.Status, UUID string) (err error) {
 	ctx, span := t.metrics.Tracer().Start(ctx, "submitter.StoreTX", trace.WithAttributes(
-		append(util.TxToAttributes(tx), attribute.String("status", status.String()))...))
+		append(txToAttributes(tx, UUID), attribute.String("status", status.String()))...))
 
 	defer func() {
 		metrics.EndSpanWithErr(span, err)
 	}()
 
 	err = t.db.PutTXS(ctx, db.TX{
+		UUID:        UUID,
 		Transaction: tx,
 		Status:      status,
 	})
@@ -274,7 +274,7 @@ func (t *txSubmitterImpl) SubmitTransaction(parentCtx context.Context, chainID *
 	}
 
 	// then we copy the transactor, this is the one we'll modify w/ no send.
-	transactor := util.CopyTransactOpts(parentTransactor)
+	transactor := copyTransactOpts(parentTransactor)
 
 	var locker mapmutex.Unlocker
 
@@ -328,7 +328,7 @@ func (t *txSubmitterImpl) SubmitTransaction(parentCtx context.Context, chainID *
 	defer locker.Unlock()
 
 	// now that we've stored the tx
-	err = t.storeTX(ctx, tx, db.Stored)
+	err = t.storeTX(ctx, tx, db.Stored, uuid.New().String())
 	if err != nil {
 		return 0, fmt.Errorf("could not store transaction: %w", err)
 	}
@@ -365,8 +365,7 @@ func (t *txSubmitterImpl) setGasPrice(ctx context.Context, client client.EVM,
 	if t.config.SupportsEIP1559(int(bigChainID.Uint64())) {
 		transactor.GasFeeCap = t.config.GetMaxGasPrice(chainID)
 
-		// only use fallback if no previous tx
-		transactor.GasTipCap, err = t.SuggestGasTipCap(ctx, client, chainID, prevTx != nil)
+		transactor.GasTipCap, err = client.SuggestGasTipCap(ctx)
 		if err != nil {
 			return fmt.Errorf("could not get gas tip cap: %w", err)
 		}
@@ -376,6 +375,7 @@ func (t *txSubmitterImpl) setGasPrice(ctx context.Context, client client.EVM,
 			return fmt.Errorf("could not get gas price: %w", err)
 		}
 	}
+	t.applyBaseGasPrice(transactor, chainID)
 
 	//nolint: nestif
 	if prevTx != nil {
@@ -398,6 +398,22 @@ func (t *txSubmitterImpl) setGasPrice(ctx context.Context, client client.EVM,
 		gas.BumpGasFees(transactor, t.config.GetGasBumpPercentage(chainID), gasBlock.BaseFee, maxPrice)
 	}
 	return nil
+}
+
+// applyBaseGasPrice applies the base gas price to the transactor if a gas price value is zero.
+func (t *txSubmitterImpl) applyBaseGasPrice(transactor *bind.TransactOpts, chainID int) {
+	if t.config.SupportsEIP1559(chainID) {
+		if transactor.GasFeeCap == nil || transactor.GasFeeCap.Cmp(big.NewInt(0)) == 0 {
+			transactor.GasFeeCap = t.config.GetBaseGasPrice(chainID)
+		}
+		if transactor.GasTipCap == nil || transactor.GasTipCap.Cmp(big.NewInt(0)) == 0 {
+			transactor.GasTipCap = t.config.GetBaseGasPrice(chainID)
+		}
+	} else {
+		if transactor.GasPrice == nil || transactor.GasPrice.Cmp(big.NewInt(0)) == 0 {
+			transactor.GasPrice = t.config.GetBaseGasPrice(chainID)
+		}
+	}
 }
 
 // getGasBlock gets the gas block for the given chain.
@@ -423,7 +439,7 @@ func (t *txSubmitterImpl) getGasBlock(ctx context.Context, chainClient client.EV
 		if ok {
 			span.AddEvent("could not get gas block; using cached value", trace.WithAttributes(
 				attribute.String("error", err.Error()),
-				attribute.String("blockNumber", util.BigPtrToString(gasBlock.Number)),
+				attribute.String("blockNumber", bigPtrToString(gasBlock.Number)),
 			))
 		} else {
 			return nil, fmt.Errorf("could not get gas block: %w", err)
@@ -470,90 +486,5 @@ func (t *txSubmitterImpl) getGasEstimate(ctx context.Context, chainClient client
 
 	return gasEstimate, nil
 }
-
-// forceNoFallbackIfZero is used to force no fallback if the tip cap is 0.
-var forceNoFallbackIfZero = false
-
-// TODO: test oracle fallback.
-func (t *txSubmitterImpl) SuggestGasTipCap(ctx context.Context, client client.EVM, chainID int, fallbackIfZero bool) (tipCap *big.Int, err error) {
-	tipCap, err = client.SuggestGasTipCap(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("could not get gas tip cap: %w", err)
-	}
-
-	// note: these ifs can be combined, but I've found the largic hard to parse if they are.
-	// if tip cap is not zero  return
-	if big.NewInt(0).Cmp(tipCap) != 0 {
-		return tipCap, nil
-	}
-
-	// past this point tip cap is 0.
-
-	// if we should force no fallback, return
-	// this is set for testing.
-	if forceNoFallbackIfZero {
-		return tipCap, nil
-	}
-
-	// if we shouldn't fallback, return
-	if !fallbackIfZero {
-		return tipCap, nil
-	}
-
-	blockHeight, err := client.BlockNumber(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("could not get block number: %w", err)
-	}
-
-	// if on the other hand, fee is 0 we should fallback to our pricer
-	oracleConfig := gas.GetConfig()
-	oracleConfig.MaxPrice = core.CopyBigInt(t.config.GetMaxGasPrice(chainID))
-
-	feeOracle := london.NewFeeOracle(wrappedLondonClient{client, chainID}, blockHeight, oracleConfig)
-	tipCap, err = feeOracle.SuggestTipCap(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("could not get tip cap: %w", err)
-	}
-	return tipCap, nil
-}
-
-type wrappedLondonClient struct {
-	client.EVM
-	chainID int
-}
-
-// ChainConfig is a fake chain config.
-// since wrapped client is only used for makeSigner
-// and on london, we just enable everything.
-func (w wrappedLondonClient) ChainConfig() *params.ChainConfig {
-	return &params.ChainConfig{
-		ChainID:                       big.NewInt(int64(w.chainID)),
-		HomesteadBlock:                big.NewInt(0),
-		DAOForkBlock:                  big.NewInt(0),
-		DAOForkSupport:                false,
-		EIP150Block:                   big.NewInt(0),
-		EIP150Hash:                    common.Hash{},
-		EIP155Block:                   big.NewInt(0),
-		EIP158Block:                   big.NewInt(0),
-		ByzantiumBlock:                big.NewInt(0),
-		ConstantinopleBlock:           big.NewInt(0),
-		PetersburgBlock:               big.NewInt(0),
-		IstanbulBlock:                 big.NewInt(0),
-		MuirGlacierBlock:              big.NewInt(0),
-		BerlinBlock:                   big.NewInt(0),
-		LondonBlock:                   big.NewInt(0),
-		ArrowGlacierBlock:             big.NewInt(0),
-		GrayGlacierBlock:              big.NewInt(0),
-		MergeNetsplitBlock:            big.NewInt(0),
-		ShanghaiBlock:                 big.NewInt(0),
-		CancunBlock:                   big.NewInt(0),
-		TerminalTotalDifficulty:       big.NewInt(0),
-		TerminalTotalDifficultyPassed: false,
-		Ethash:                        nil,
-		Clique:                        nil,
-	}
-}
-
-var _ backend.OracleBackendChain = &wrappedLondonClient{}
 
 var _ TransactionSubmitter = &txSubmitterImpl{}