+++ packages/explorer-ui/pages/tx/[kappa].tsx
@@ -5,7 +5,7 @@ import { Error } from '@components/Error'
 import { StandardPageContainer } from '@components/layouts/StandardPageContainer'
 import { useRouter } from 'next/router'
 import { useSearchParams } from 'next/navigation'
-import { CHAIN_EXPLORER_URLS, BRIDGE_CONTRACTS, CCTP_CONTRACTS } from '@constants/networks'
+import { BRIDGE_CONTRACTS, CCTP_CONTRACTS } from '@constants/chains/index'
 import { GET_BRIDGE_TRANSACTIONS_QUERY } from '@graphql/queries'
 import { API_URL } from '@graphql'
 import { HorizontalDivider } from '@components/misc/HorizontalDivider'
@@ -14,6 +14,7 @@ import { formatDateTime } from '@utils/formatDate'
 import CopyTitle from '@components/misc/CopyTitle'
 import { IconAndAmount } from '@components/misc/IconAndAmount'
 import { BridgeTransactionTable } from '@components/BridgeTransaction/BridgeTransactionTable'
+import { CHAINS_BY_ID } from '../../constants/chains'
 
 const link = new HttpLink({
   uri: API_URL,
@@ -41,9 +42,9 @@ export default function BridgeTransaction({ queryResult }) {
   }
   const generateBridgeAddress = (chainID, eventType) => {
     if (eventType == 10 || eventType == 11) {
-      return CHAIN_EXPLORER_URLS[chainID] + '/address/' + CCTP_CONTRACTS[chainID]
+      return CHAINS_BY_ID[chainID].explorerUrl + '/address/' + CCTP_CONTRACTS[chainID]
     }
-    return CHAIN_EXPLORER_URLS[chainID] + '/address/' + BRIDGE_CONTRACTS[chainID]
+    return CHAINS_BY_ID[chainID].explorerUrl + '/address/' + BRIDGE_CONTRACTS[chainID]
   }
   const transaction = queryResult.bridgeTransactions[0]
   const { pending, fromInfo, toInfo } = transaction
@@ -147,7 +148,7 @@ export default function BridgeTransaction({ queryResult }) {
                     rel="noreferrer"
                     className="text-white break-all text-sm underline"
                     href={
-                      CHAIN_EXPLORER_URLS[fromInfo.chainID] +
+                      CHAINS_BY_ID[fromInfo.chainId]?.explorerUrl +
                       '/tx/' +
                       fromInfo.hash
                     }
@@ -220,7 +221,7 @@ export default function BridgeTransaction({ queryResult }) {
                       rel="noreferrer"
                       className="text-white break-all text-sm underline"
                       href={
-                        CHAIN_EXPLORER_URLS[toInfo.chainID] +
+                        CHAINS_BY_ID[toInfo.chainID]?.explorerUrl +
                         '/tx/' +
                         toInfo.hash
                       }