+++ packages/synapse-interface/components/StateManagedBridge/ToChainListOverlay.tsx
@@ -31,6 +31,7 @@ export const ToChainListOverlay = () => {
     .pickBy((value) => _.includes(toChainIds, value.id))
     .values()
     .value()
+    .filter((chain) => !PAUSED_TO_CHAIN_IDS.includes(chain.id))
 
   possibleChains = sortChains(possibleChains)